Restructured files for pip packaging
