+++ MANIFEST.in
@@ -1,0 +1,1 @@
+recursive-include cmu_course_api/data *
