+++ cmu_course_api/__init__.py
@@ -1,0 +1,1 @@
+from .aggregate import get_course_data
