+++ cmu_course_api/aggregate.py
@@ -1,0 +1,64 @@
+#!/usr/bin/env python3
+# @file course-api.py
+# @brief One script to rule them all.
+#
+#        Downloads schedule data for a specific semester, including course
+#        meeting times, course descriptions, pre/corequisites, FCEs, and so on.
+#        Output is parsed into a single JSON output file.
+#
+#        If you want to download just one of Course description data, course
+#        schedule data, or FCE data see the scripts/ folder.
+#
+# @author Justin Gallagher (jrgallag@andrew.cmu.edu)
+# @since 2015-04-07
+
+
+import json
+import os.path
+from cmu_course_api.parse_descs import parse_descs
+from cmu_course_api.parse_schedules import parse_schedules
+from cmu_course_api.parse_fces import parse_fces
+
+
+# Constants
+DESC_SOURCES = os.path.join(os.path.dirname(__file__), 'data/schedule_pages.txt')
+
+
+# @function aggregate
+# @brief Combines the course descriptions, schedules, and FCEs data sets into
+#        one object.
+# @param descs: Course desciptions object as returned by parse_descs.
+# @param schedules: Course schedules object as returned by parse_descs.
+# @param fces: FCEs object as returned by parse_descs.
+# @return An object containing the aggregate of the three datasets.
+def aggregate(descs, schedules, fces):
+    courses = {}
+
+    for department in schedules:
+        for course in department['courses']:
+            for desc in descs:
+                if ('num' in desc and desc['num'] == int(course['num'])):
+                    desc['department'] = department['department']
+                    desc['lectures'] = course['lectures']
+
+                    num = desc['num']
+                    del desc['num']
+
+                    courses[str(num)] = desc
+
+    return {'courses': courses, 'fces': fces}
+
+
+# @function course_api
+# @brief Used for retrieving all information from the course-api for a given
+#        semester.
+# @param semester: The semester to get data for. Must be one of S, M1, M2, or F.
+# @param username: Username to use for andrew authentication.
+# @param password: Password to use for andrew authentication.
+# @return Object containing all course-api data - see README.md for more
+#        information.
+def get_course_data(semester, username, password):
+    descs = parse_descs(DESC_SOURCES)
+    schedules = parse_schedules(semester)
+    fces = parse_fces(username, password)
+    return aggregate(descs, schedules, fces)
