+++ cmu_course_api/data/schedule_pages.txt
