+++ cmu_course_api/parse_descs.py
