+++ cmu_course_api/parse_schedules.py
