--- requirements.txt
@@ -1,2 +1,0 @@
-beautifulsoup4==4.4.1
-cmu_auth==0.1.6
